{"schema":"decodex/commit/1","summary":"Modularize capture host material views","authority":"manual"}
